+++ CHANGELOG.md
@@ -4,6 +4,7 @@ NEXT_VERSION_BUMP: PATCH
 ## XX.XX.XX - 20XX-XX-XX
 
 * [Added] Naira card support in PaymentSheet.
+* [Added] Naira bank transfer support in PaymentSheet.
 * [Added] GoPay support in PaymentSheet.
 * [Added] MoMo support in PaymentSheet.
 * [Added] GCash support in PaymentSheet.